cselectiveAPI-main/requirements.txt
django
djangorestframework
environs
drf-yasg
